requirements.txt
asgiref==3.3.1
Django==3.1.4
django-environ==0.4.5
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2020.4
PyYAML==5.3.1
sqlparse==0.4.1
uritemplate==3.0.1
